avoid needless subproject rebuilds by passing 'stable' includes
The trick is to pass CPPFLAGS prefer includes from sources, rather
than from installed headers. Because make install always updates
timestamps of installed headers causing needless rebuilds of depended
subprojects.

This avoids most of rebuilds. The only exception is few files in moxi
that depend on libconflate, for which it's hard to pass -I<to-source>.

Change-Id: Ib0a48c26a54af5948058cf4540bedd063ebd4ec0
Reviewed-on: http://review.membase.org/5200
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Steve Yen <steve.yen@gmail.com>