automagically update top-level Makefile
Because we depend on GNU Make we can use it's nice features. One of
them is automatic update of Makefile if there are rules for that. So
we just define rule for it.

This fixes one of annoyances of repo which copies tlm/Makefile.top
into top-level Makefile, so updates of tlm/Makefile.top do not apply
to top-level Makefile.

Change-Id: I105fdfb95fcb11fd5f96b1a2a6483d37f0808223
Reviewed-on: http://review.couchbase.org/11184
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Trond Norbye <trond.norbye@gmail.com>