prevent autotools stuff rebuilding for snappy
git may checkout files in arbitrary order. Looks like on GNU/Linux
with high precision mtime it can cause produced Makefile to think that
Makefile.in or config.h.in or even aclocal.m4 need to be rebuilt. And
that fails on my machine because it has different versions of
autotools.

The cure is to touch this files in correct order.

Change-Id: I1eacdc606c3885cdb449c44fbe90c6ee0b914a59
Reviewed-on: http://review.couchbase.org/14006
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>