Add include to the platform includes
The platform library does some runtime checks to determine
the availability of "standard" functions. The knowledge about
their existence is needed at compile time for other modules
as well, and in order to support "out of source builds" the
file containing the checks is stored in the "build directory"
for the platform library. To avoid updating the include path
of _all_ individual projects lets add that path globally.

Change-Id: I8abba86aef35daba3159d66b9f79d8094f2f83ae
Reviewed-on: http://review.couchbase.org/42382
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: Trond Norbye <trond.norbye@gmail.com>