MB-12711: Make CMake pick up the right Erlang include path
There was a hack introduced for Debian. Though it seems that it was
a Debian bug. With a more recent version of Debian testing, it works
without the hack.

This commit is basically a revert of
565869a9c9929382cc41093c947544345e40e40e [1].

The problem with the hack was, that it tried to check for a header file
(erl_native_features_config.h) that only exists in Erlang >=17.0. When you
tried to compile Couchbase with a local Erlang <17.0, but had a system wide
installation of Erlang >=17.0, it would pick up the system wide installed
header files.

[1]: https://github.com/couchbase/tlm/commit/565869a9c9929382cc41093c947544345e40e40e

Change-Id: I5390b80404bd978cba807cf27d64ae143541de16
Reviewed-on: http://review.couchbase.org/43479
Tested-by: Volker Mische <volker.mische@gmail.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>