MB-12745: Check for usable jemalloc library
When using jemalloc as the memory allocator, we require a version
which has it's symbols prefixed (i.e. named je_malloc, je_free etc. as
opposed to malloc, free) as memcached needs to be able to interpose
it's own version of malloc et al which will in turn explictly call
je_XXX.

Therefore add a check to the CMake script to only set JEMALLOC_FOUND
if library of that configuration is found.

Change-Id: I6801e508efd607d4f8ba5283c95cfdf84683f63d
Reviewed-on: http://review.couchbase.org/43559
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: David Haikney <david@haikney.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>