cbdeps: Ensure that a new dep version is detected
cbdeps currently checks for the existance of a directory named
${dep_name} to see if a dependancy has already been extracted. However
as the directory name doesn't include it's version, if the dependancy
manifest is updated with a new version then cbdeps fails to notice and
hence the new version is not downloaded.

Fix this by recording the extracted version, and checking this against
the required version, ensuring that a different version will trigger a
re-extract.

Change-Id: If07c2e874321b81e0cd239c907acfc86c9a02e1f
Reviewed-on: http://review.couchbase.org/43644
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Reviewed-by: Mark Woosey <mark.woosey@couchbase.com>
Reviewed-by: Adam Taylor <adamt@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>