MB-10496: Add jemalloc as a cbdeps dependancy for CentOS 6
This will cause jemalloc to be the preferred memory allocator for
CentOS 6.

Note that the previous version (cb1) of the jemalloc cbdeps package
was incorrectly uploaded as cb2. As a consequence, Centos6 (only) uses
version cb3. This is identical to cb2 for the other platforms, and is
what cb2 /should/ have been for Centos 6 :(

Change-Id: I0f2d5fd4323c268013a8a23f54c80c2a2c70639f
Reviewed-on: http://review.couchbase.org/44111
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>