Don't strip the platform from the filename
This patch preserves the full filename for the file as it
is stored in cbdeps when the file is put into the .cbdepcache
directory. This makes it easier to try out new dependencies
since you don't have to remember to rename the file when
you drop it into the cache directory.

Change-Id: I561f87db18ad8097dbb76433de4e482505c3f8d6
Reviewed-on: http://review.couchbase.org/44241
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: Trond Norbye <trond.norbye@gmail.com>