Only mark selected cbdeps headers as system headers
Some cbdeps headers (e.g. breakpad) have a large number of warnings in
them, and so we mark their directories as system directories (as
warnings in system headers are ignored.

However a side-effect of this is it changes the compiler search order
- non-system paths (-I) are searched before system paths. Therefore if
a header file exists both in a standard system location
(e.g. /usr/local/include) and in one of our paths then adding to
CB_SYSTEM_HEADER_DIRS may result in the compiler picking up the wrong
version.

An example of this (and the motivation for this change) is V8 -
historically the user had to provide this themself and so it was
common to use homebrew on OS X to install it, which places v8.h in
/usr/local/include. However we recently packaged V8 using cbdeps,
which has a more up-to-date version of the header (and is incompatible
with the older one). As a result of marking cbdeps' include directory
(containing v8.h) as a system directory, the compiler would end up
using the older v8.h from /usr/local/include.

To fix this, we only add headers which (1) have known warning issues
and (2) are unlikely to exist in a normal system location. That set is
currently breakpad.

Change-Id: Ia0a7390c12e79749d0c753373369a3b5f6af5a69
Reviewed-on: http://review.couchbase.org/44502
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>