MB-13564: Properly locate the libraries reported by icu-config
Normally icu-config would just print out -L/path -llibname
causing that being passed to the compiler when we tried to
link with libicu. Normally that would be sufficient, but
unfortunately it'll cause the binary to fail to locate the
libraries at runtime if they're installed in a non-standard
location. CMake would do the proper magic inserting
-Wl,-rpath,path if we use FIND_LIBRARY to detect the libraries.

Change-Id: I0549d7f06f9769abe130302e04c22647c4d0d256
Reviewed-on: http://review.couchbase.org/47157
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Volker Mische <volker.mische@gmail.com>