MB-13722: RelWithDebInfo optimization flags should match Release
Adjust the C & C++ compiler flags for Release-with-debug-info builds to
use the same optimization level as a Release build, the only difference
being if debugging information (.PDB files) are created during compilation.

This allows us to use RelWithDebInfo in the production "Release" build and
have no reduction in code optimization level (and hence performance).

Change-Id: I2f59b74cf9b9bac4e2d2e9c662220cc2a67b0d41
Reviewed-on: http://review.couchbase.org/47626
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Dave Rigby <daver@couchbase.com>