MB-13723: [GCC/Clang] RelWithDebInfo optimization flags should match Release
Adjust the GCC/Clang C & C++ compiler flags for
Release-with-debug-info builds to use the same optimization level as a
Release build, the only difference being if DWARF debugging
information is created during compilation.

This allows us to use RelWithDebInfo in the production "Release" build and
have no reduction in code optimization level (and hence performance).

Change-Id: Iec3dc6837e82567fb15d5a57d86fdd96a7fb1325
Reviewed-on: http://review.couchbase.org/47686
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>