MB-13753: Add $ORIGIN/../lib to Breakpad runtime linker path
We build a helper program - minidump-2-core - as part of the Breakpad
cbdep. On Ubuntu 12.04 we use a newer C++ compiler (and hence new
libstdc++.so) than the default. We ship this newer libstdc++.so with
the product (in /opt/couchbase/lib); but minidump-2-core doens't have
this path in it's runtime linker search path, so it fails to start
with the following error if installed on a vanilla Ubuntu 12.04
system:

    minidump-2-core: /usr/lib/x86_64-linux-gnu/libstdc++.so.6: version `GLIBCXX_3.4.20\' not found

Fix this by adding $ORIGIN/../lib (i.e. /opt/couchbase/lib) to the
search path.

Change-Id: I069e0ec693a24c3db5eb5a3ccddbc53d2f73ed6e
Reviewed-on: http://review.couchbase.org/47853
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>