CBDeps always extracts dep tar file even when already extracted
The CMake macro to check if a cbdep has already been extracted checks
for the wrong variable name (EXISTING_VERSION instead of
EXPLODED_VERSION), hence it always extracts the tar.gz file.

Fix this by checking for the correct variable. Additionally, abort the
build if the extract fails, so we don't write the stamp file and
potentially think we have previously extracted when we haven't.

Change-Id: I1cfce55d96020344f3d389f51d2de96440a33b91
Reviewed-on: http://review.couchbase.org/49448
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>