Valgrind: Suppress leak in libstdc++ pool constructor
As part of a recent update to the Docker image, libstdc++ allocates
72K of memory which is still reachable at the end of the program
(i.e. not freed). This isn't strictly a 'leak' - it's a memory pool
(https://gcc.gnu.org/bugzilla/show_bug.cgi?id=65434) but if Valgrind
is run with --show-leak-kinds=all then it is reported as an error.

As this is code out of our control and to ensure this doesn't cause
memory leak failure in our unit tests, add a Valgrind suppressions
file containing this specific error (so we can still detect other
reachable memory issues).

Change-Id: I983f8c90d1a236566ceb54dda29e70d491a82378
Reviewed-on: http://review.couchbase.org/51230
Tested-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: buildbot <build@couchbase.com>