Rebuild jemalloc with Valgrind support
Rebuild jemalloc with support for Valgrind, by ensuring that jemalloc
can locate Valgrind's headers at build time. This allows jemalloc to
detect at runtime if it's running under Valgrind; and consequently
notify Valgrind about it's heap usage.

This means Valgrind has a correct view of jemalloc's memory
usage and means more memory issues / leaks can be detected.

NOTE: This requires that the actual OS we build the jemalloc cbdep on
has Valgrind installed. This is currently the case for Ubuntu12.04[1],
which is the platform currently used for CV and hence this will allow
us to spot leaks accurately there.

[1]: https://github.com/couchbase/build/blob/master/docker/buildslaves/ubuntu-12.04/jenkins/Dockerfile

Change-Id: I95657c0b302c9a83c7adfaea660d5dc1b9f9c0a4
Reviewed-on: http://review.couchbase.org/52541
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Chris Hillery <ceej@couchbase.com>