Valgrind: Add suppressions for memcached thread 'reachable' allocations
Add Valgrind suppressions for allocations from memcached background
threads which are not cleanly 'joined' on shutdown; hence they have
some allocations which are still reachable. Given there's no easy way
to signal to these threads that the process has terminated (and they
are not actual leaks) we choose suppress them.

Change-Id: I5577ab0b651201ab17dbbd67b794bf400e4601c4
Reviewed-on: http://review.couchbase.org/52660
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>