MB-15435: Pass ThreadSanitizer linker flags to CGo
When building with ThreadSanitizer, we need to ensure that any CGo
Targets (e.g. 2i indexer) have the correct linker flags otherwise they
will fail to link due to missing symbols:

    /home/couchbase/couchbase/build/forestdb/libforestdb.so: undefined reference to `__tsan_read8'
    /home/couchbase/couchbase/build/forestdb/libforestdb.so: undefined reference to `__tsan_read1'
    /home/couchbase/couchbase/build/forestdb/libforestdb.so: undefined reference to `__tsan_init'
    ...

To solve this ensure that CGO_LDFLAGS includes the appropriate
ThreadSanitizer linker flags; and we use the selected CMake compiler for
CGo.

Change-Id: I79b794e8ca0b3706708924e2b6d7e163c2581a4c
Reviewed-on: http://review.couchbase.org/53020
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>