Add CB_{C,CXX}_FLAGS_NO_OPTIMIZE variable
Sometimes it is desirable to disable optimization for certain targets;
however on Windows it isn't possible to simply use 'Debug' build flags
on some object files and 'Release' flags on others if they are linked
together into a single executable, due to the different incompatible
CRTs.

The initial solution to this (see dfc6abe) was to override the default
Debug flags to always use the 'Release' CRT. While this solves the
problem of linking Debug and non-Debug objects, it essentially
prevents the user from building a "true" Debug build on Windows with
the Debug CRT.

Therefore sovle this in a different way. Add a two new variables:
CB_{C,CXX}_FLAGS_NO_OPTIMIZE which can be added to a target to /only/
disable optimization, and to leave all other Release/Debug flags
as-is.

Once these variables are used by all targets which want to build
without optimizations, the overrides to CMAKE_{C,CXX}_FLAGS_DEBUG can
be removed (in a follow-up patch).

Change-Id: I5b6fca52b56c26cc7aa01a28e4f194f8adbfb694
Reviewed-on: http://review.couchbase.org/55467
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>