Enable color diagnostics when using Clang with Ninja
When using Ninja as the build generator, Clang fails to auto-detect
that it is running on a terminal (and hence doesn't print color
diagnostics).

Solve this by explcitly telling Clang to use color when Ninja is the
build generator.

Note that Ninja is smart enough to strip out any color escape
sequences if stdout isn't a terminal, so this is safe when logging
Ninja's output to a file.

Change-Id: I951348d104dcf66431f6c82665b7cee47c2926fe
Reviewed-on: http://review.couchbase.org/55718
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: buildbot <build@couchbase.com>