Update valgrind suppression for parent monitor thread
We don't have any shutdown logic for the parent monitor thread
(and it is only used in testing to make sure that memcached
terminates if testapp dies).

We've changed the way we allocate and release the Thread Context,
and in order to avoid having to create another suppression fix
the next time we change this I'm just suppressing anything
allocated after the cb_create_named_thread.

Change-Id: I6b9e5b45921482c9debca35a8c65cf1ff4eb2acd
Reviewed-on: http://review.couchbase.org/56268
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>