MB-16702: Update cbdeps jemalloc to v4.0.4 [1/2]
We are currently building a development version of jemalloc between
3.6.0 and 4.0 fir sherlock - we needed this version to support some of
the features we needed for memory defrag.  Given that jemalloc has now
released version 4.0 (i.e. the stable, cleaned up version of what we
are already using), we should upgrade to it for Watson.

Note: as part of this we no longer install jemalloc's header files
(under install/include/jemalloc) - these are not needed as part of the
shipped product. Instead applications should pick them up from
build/tlm/deps/jemalloc.exploded

Change-Id: I423fb8067e626fe451073563bf92ec0009950424
Reviewed-on: http://review.couchbase.org/56634
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>