CBD-1751: Support multiple versions of Go
Introduces "multi-go" mode and makes it the default. In this mode, all
GoInstall() targets may declare a GOVERSION (the default is set to 1.5.2
currently). All desired Go compilers will be downloaded as part of the
build and cached locally in a manner similar to cbdeps.

The go compiler caches compiled go libraries in pkg/ directories. To
prevent collisions, each go version is given a separate pkg/ directory
in the build tree using the -pkgdir go option. Because Go 1.4.x did not
support this option (meaning 1.4 projects will all share the default
pkg/ directory), the build is limited to a single 1.4.x version of Go,
which is defined by the variable GO_14x_VERSION. Targets may specify
"GOVERSION 1.4.x" to use this version. Attempting to specify an explicit
version such as "GOVERSION 1.4.2" will result in a CMake error.

Change-Id: If82c3b78bd6774d408f2a3244d4d5bf2f8f60131
Reviewed-on: http://review.couchbase.org/59245
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>