MB-18591: Use snapshots in V8
A snapshot is a dump of the V8 heap. If it is enabled, a default
snapshot is created when V8 is compiled and then bundled with V8.
This makes the startup significantly faster.

We were compiling V8 without snapshot support. This means that on
every startup and context creation the default JavaScript code V8 contains
needed to be compiled. This slowed things down a lot.

Change-Id: I819aeab92328013de40434ef83a8ff697d0f0e13
Reviewed-on: http://review.couchbase.org/61224
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Reviewed-by: Harsha H S <hhs.couchbase@gmail.com>
Tested-by: Volker Mische <volker.mische@gmail.com>