MB-19203: Allow use of non-windows cbdeps from Sherlock
Add the cbdeps from Sherlock (for non-Windows platforms) to cbdeps
manifest. The main motivation for this is to be able to build 3.0.x on
the current (Sherlock/Watson) commit-validation images, for example to
run ThreadSanitizer on 3.x

Note that cbdeps is *not* enabled by default on tlm/3.0, so unless the
user explicitly turns them on (via -DCB_DOWNLOAD_DEPS=1) then this
patch has no effect.

Change-Id: I30884a6cadecf3c305dd950e07de31e552babfd9
Reviewed-on: http://review.couchbase.org/62773
Tested-by: buildbot <build@couchbase.com>
Well-Formed: buildbot <build@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>