Mark googletest as EXCLUDE_FROM_ALL so it isn't installed
We shouldn't run the 'install' target for googletest, as we don't want
to actually ship (copy to install/include) all it's headers. Therefore
mark the subdirectory as EXCLUDE_FROM_ALL, so only the targets we
explicitly invoke (e.g. building the library so we can link to it)
will be invoked.

Change-Id: I7046692607ca2c0e442d2d3fec2afab67eecb93c
Reviewed-on: http://review.couchbase.org/63714
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Will Gardner <will.gardner@couchbase.com>
Reviewed-by: Hari Kodungallur <hari.kodungallur@couchbase.com>