MB-20036: on windows use erlang built using VS2010
VS2013 has a FMA3 bug that caused the erl.exe (happens only on certain
cpu settings). The bug doesn't exist on any other version of visual studio.
The windows erlang build we were distributing so far (downloaded from
erlang), was built with VS2010 and so we build erlang with VS2010 also

Change-Id: I01033dd1dba4852ffe0975f956d4e54df4a04ec4
Reviewed-on: http://review.couchbase.org/65645
Tested-by: Hari Kodungallur <hari.kodungallur@couchbase.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>