Use the real (base) version of erlang in our version
We forked off erlang from R16B03-1 and we should be using that
instead of just R16B03. Windows already started using R16B03-1.
Windows already has the version R16B03-1-couchbase-cb1. And from
the next build we should have both win and linux at
R16B03-1-couchbase-cb2

This is going to be really confusing since the current version
on linux is R16B03-couchbase-cb2. The next one will be
R16B03-1-couchbase-cb2. That means the cb-version is not changing,
but the erlang version is updated to its original value. We are
not changing the cb version because the windows version is
currently at R16B03-1-couchbase-cb1. So to get both platforms
in sync, we are going to use R16B03-1-couchbase-cb2 for the
next build.

Change-Id: If987924fac76aab181537b642a0121015bd9cc50
Reviewed-on: http://review.couchbase.org/66591
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Hari Kodungallur <hari.kodungallur@couchbase.com>