Add ability to disable compilation of phosphor trace macros
This patch allows for global disabling of phosphor trace macros
by use of the cmake flag 'PHOSPHOR_DISABLE'. This flag is (for now)
ON by default and can be set to OFF like so:

    cmake .. -DPHOSPHOR_DISABLE=OFF

It is also OFF by default if using the
COUCHBASE_KV_COMMIT_VALIDATION flag so that the Phosphor trace
macros are at least compiled in for commit validation of KV-Engine
components.

The purpose of this patch is to allow for inclusion of Phosphor in
KV-Engine components short-term without having to worry about
impacting Perf/QE testing.

The patch works by setting the PHOSPHOR_DISABLED compile definition
and by setting the optional PHOSPHOR_LIB cmake variable to the
empty string (To allow for optional linkage of phosphor itself).

Change-Id: I88c98a2d4f4f0a1af7c38fbf22dcbb1fb706011e
Reviewed-on: http://review.couchbase.org/66645
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>