clang-format: Remove configurations not supported by v3.8
Remove a (small) number of configuration parameters which are not
supported by Clang-Format v3.8. This allows our config file to work
(otherwise it's rejected) with clang 3.8, which is the version
shipping on Ubuntu 16.04.

Note of the 5 removed fields, 3 are Javascript-related (to irrelevent
to us), and the other two (BreakStringLiterals and IncludeIsMainRegex)
are relativley minor and shouldn't make much of a difference.

Change-Id: I782a9fdb99fbf5819a1d93683159847f960d38ae
Reviewed-on: http://review.couchbase.org/70074
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>