Valgrind: Add support v3.12 (disable --soname-synonyms)
Valgrind 3.12 changes the default value for the --soname-synonyms
option - from the release notes [1]:

    Replacement/wrapping of malloc/new related functions is now done
    not just for system libraries by default, but for any globally
    defined malloc/new related function (both in shared libraries and
    statically linked alternative malloc implementations).  The
    dynamic (runtime) linker is excluded, though.

    To only intercept malloc/new related functions in system libraries
    use --soname-synonyms=somalloc=nouserintercepts (where
    "nouserintercepts" can be any non-existing library name).

This change in defaults causes a number of ep-engine unit tests to
fail; as it looks like Valgrind is too aggressive in hooking our
various malloc functions (some combination of jemalloc / cb_malloc).

Given that the memory allocator we use (jemalloc) already registers
itself with Valgrind correctly, disable this new feature as per the
release notes.

[1]: http://valgrind.org/docs/manual/dist.news.html

Change-Id: Ic73f2729e9aa6ba19aa938e828514b0f5e16ad3b
Reviewed-on: http://review.couchbase.org/71991
Tested-by: buildbot <build@couchbase.com>
Tested-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>