MB-22191: Always set PRODUCT_VERSION reasonably
Make PRODUCT_VERSION a cached variable to avoid losing it between builds.
If it is set to "" (the default), override the cached value by attempting
to introspect the VERSION from the manifest plus a build number of 0000.
If introspection fails, ultimate default value is "0.0.0-9999".

Since we produce a better default, we removed the ability to easily
override it via Makefile. In general it's bad to automatically pass -D
arguments to CMake with default values because CMake assumes those are
the most correct values.

Change-Id: I209456870c900cdd392ab010cb03e3512a78da27
Reviewed-on: http://review.couchbase.org/77692
Reviewed-by: Kenneth Lareau <kenneth.lareau@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>