Require at least MSVC 2015 for C++11 support
While MSVC 2013 has okay C++11 support, it also has many bugs (grep
the codebase for 'MSVC 2013' to see all the workarounds we've had to
do).

Given we are now building with MSVC 2015, make this 'official' by
declaring that we need at least MSVC 2015 for C++11 support.

In effect this means the build will explicilty fail early if you
attempt to build with an older version.

Change-Id: I10ebf9e0e4a5b663d6f7eb1c0675367b2d46bc28
Reviewed-on: http://review.couchbase.org/78960
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>