MB-26221 [1/2] [RocksDB] Enable Snappy compression
RocksDB uses Snappy compression by default if Snappy is linked to the
RocksDB library. With this change we link Snappy when we build RocksDB.

Note that rocksdb is now the second cbdep package to require Snappy as
a dependancy (after python-snappy). As such, it is necessary to bypass
the normal check of only downloading and installing a dep once; as
both rocksdb and python-snappy need to have snappy downloaded and
extracted into their respective build directories.

Change-Id: I8b59f7a39003f8b4735876f5a3045da5ff63bf9c
Reviewed-on: http://review.couchbase.org/83957
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Dave Rigby <daver@couchbase.com>