MB-26295 Add CB_NON_PRODUCTION_ASSERTS CMake option, initially OFF
It is desirable to add asserts in the codebase for functionality which
is non-critical - while we want to expose bugs early in development /
testing (via an asset/exception firing), we don't want to trigger such
asserts in production.  For example, asserting that statistical
counters are of an expected range (don't underflow) - we would like to
detect such bugs; but they arn't serious enough to warrant crashing in
a customer environment.

As such, add a new CMake option - CB_DEVELOPMENT_ASSERTS - which can
be used to guard such checks. This will be enabled during the
development phase of a release; but disabled before Beta/Pre-Release.

Change-Id: Ide87c48e6ff54ddcb6fd696e4c5464c9ae62f551
Reviewed-on: http://review.couchbase.org/84095
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>