MB-26863 [1/2] Add third_party/CMakeLists.txt for gsl-lite
Add a CMakeList for the new third_party/gsl-lite module.

Note this patch just adds the CMakeFile - it isn't used/referenced
anywhere yet. This is necessary as we need to have the file available
before the change to the manifest is made to actually checkout
gsl-lite into the correct third_party/ directory.

Change-Id: Ib02b296cb98b75635ee47b6276ef6170b860237b
Reviewed-on: http://review.couchbase.org/85647
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Sriram Ganesan <sriram@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>