MB-26298: RocksDB cbdep: Upgrade to v5.8 for FastCRC32 fix
There has been a number of issues in getting SSE4.2 accelrated CRC32c
(aka Fast CRC32) working on all platforms - even with the last tlm
change to set USE_SSE=1 during the Rocksdb build it still wasn't
enabled on some Linux platforms (including CentOS 7).

The latest is in the GNU Make build system for RocksDB - while it
correctly adds the correct '-msse4.2' compiler when testing SSE4.2
support in the compiler, it fails to set the C++ standard flag
(-std=c++11), which causes the test program to fail and hence never
sets the HAVE_SSE42 #define.

This is fixed in v5.8[1], so upgrade to RocksDB to that version to
solve this issue.

Note that in 5.8 RocksDB builds without RTTI by default. We still need
RTTI as we build the rest of KV-Engine with it, so need to explicitly
enable it now.

[1]: https://github.com/facebook/rocksdb/commit/c5f0c6cc660f1f4a8051db2aac3b8afc17818e70

Change-Id: Ibc447ab9c1b594612dfdfbbfa28253bd89c3574e
Reviewed-on: http://review.couchbase.org/85319
Reviewed-by: Kenneth Lareau <kenneth.lareau@couchbase.com>
Tested-by: Dave Rigby <daver@couchbase.com>