MB-26684 [RocksDB] Fix SegFault when running unit tests for MemCheck
The new features we introduced in RocksDBKVStore led to the execution of
the 'malloc_usable_size' function. While the 'new' and 'delete' operators
in 'libstd++' are correctly redirected to the Valgrind ones, we need
to tell Valgrind to redirect also the 'malloc_usable_size' function in
'libplatform'. Missing that leads to SegFault when the non-Valgrind
'malloc_usable_size' is called giving in input a pointer to a
Valgrind-allocated block of memory.

Change-Id: If0f2a8c936319abbbd8aacebf8c32c95107a414e
Reviewed-on: http://review.couchbase.org/85320
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>