MB-26800: cbdeps: always use 'lib' for LZ4 LIBDIR
LZ4's CMakeList.txt installs the compiled libraries to the value of
CMAKE_INSTALL_LIBDIR (via GNUInstallDirs module). On non-Debian based
Linux systems, this is 'lib64/' for x86-64. This is not what our
packaging step expects, so it results in the cbdep install script
failing as it looks in the wrong directory.

Fix by explicitly setting CMAKE_INSTALL_LIBDIR to always be 'lib'.

Change-Id: Ibeda190cef5756fdccc6c09cb7eea5dd14a5dd97
Reviewed-on: http://review.couchbase.org/85524
Tested-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>