MB-27673: Add new CMake function to check if unit tests should be built
Define a function (if not already defined) which checks if the given
module name exists in the list of modules to disable unit tests for.

The list of modules must be specified in either a cmake variable or
in an environment variable. The name and format of the variable
is the same: COUCHBASE_DISABLED_UNIT_TESTS and the format is:
module1;module2;module3.

Ex:

    export COUCHBASE_DISABLED_UNIT_TESTS="kv_engine;platform"

The intended use in the projects looks like this:

    check_unit_test_enabled(kv_engine COUCHBASE_BUILD_KV_UNIT_TESTS)
    if (COUCHBASE_BUILD_KV_UNIT_TESTS)
        ... the user wants the unit tests to be built ...
    endif()

Change-Id: Ifc38600acb326ea8e2c98d5e76c3673be833d4d9
Reviewed-on: http://review.couchbase.org/88352
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>