MB-28749: Import GoogleTest modules from CMake v3.10.3
Add the GoogleTest modules from CMake v3.10.3[1] to tlm; renamed to
add the 'Couchbase' prefix to them.

These will be used to auto-register GoogleTests with CMake / CTest for
KV-Engine, however we need to make modifications to the original code
(coming in a subsequent patch) to support adding whole GoogleTest
suites instead of individual tests.

[1]: https://github.com/Kitware/CMake/blob/v3.10.3/Modules/GoogleTest.cmake

Change-Id: Icf683b948c5f3e5a2654314ee324b234dc7dd43f
Reviewed-on: http://review.couchbase.org/91257
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Tim Bradgate <tim.bradgate@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>