Define COUCHBASE_ENTERPRISE_EDITION compiler macro
Define a symbol we might check at compile time to take different
actions for community / enterprise edition.

The current need is for the audit daemon to be able to ignore
audit descriptors only available in enterprise edition. It could
ignore missing files to achieve the same, but then we would not
detect the case where the file is missing (but expected to be
there).

Change-Id: Ib9881fab8b9da839e4df7b0b860222f07a554bc5
Reviewed-on: http://review.couchbase.org/92607
Reviewed-by: Tim Bradgate <tim.bradgate@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>