MB-29347: Change Watson Mac OS builds to later version of Erlang
This change updates the Erlang used in Watson builds to
R16B03-1-couchbase-cb8 for Mac OS builds. This Erlang includes
fixes for:
* MB-24741: crypto problem on High Sierra
* MB-27322: fix for Bleichenbacher attack

With folks in the development team working on High Sierra it's convenient to
have a build that works. And if we're upgrading Erlang, it seems wise to include
the small fix for the Bleichenbacher attack.

Change-Id: I75bbb0f29249c939f8195239835bf9f7e6f97440
Reviewed-on: http://review.couchbase.org/93009
Tested-by: Dave Finlay <dave.finlay@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>