UBSan: disable recovery upon error
By default UBSan just prints detected errors to stdout; and continues
running. While this is useful when using it as part of a production
build; it makes it harder to spot any bugs it finds (as one has to
carefully check any output from the test).

Given we use UBSan primarily during commit-validation, change this
behaviour to not recover, and instead exit with a non-zero status
code.

Change-Id: I102eb6a70e2b784c6104d05f29e08642f64a1a71
Reviewed-on: http://review.couchbase.org/95684
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Reviewed-by: Tim Bradgate <tim.bradgate@couchbase.com>