Add support to skip the install step of the cbdeps
When calling DECLARE_DEP() it automatically tries
to do add_subdirectory() on the package if the tgz file
includes a CMakeLists.txt. This makes it "hard" to use
when building other dependencies, as it may not have
set CMAKE_INSTALL_PREFIX, but wants to use the
files directly from the exploaded archive.

Change-Id: I58ff64062319ef83c0fc1a9bf2f2c75e4db34f3f
Reviewed-on: http://review.couchbase.org/93869
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>