Remove support for Sun Studio compilers
We've not build Couchbase with these compilers since back
in the 2.x / early 3.x generations and we've moved to require
a newer C++ standard since that so I highly doubt it'll
compile.

By keeping the files in there and set some flags for the compiler
we might give a false indication that we've actually tested this.

Change-Id: Id5f861919c60b5ca8cea86493d1900890876713d
Reviewed-on: http://review.couchbase.org/94752
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Owen <owend@couchbase.com>