Move forward to CMake policy 3.2
When declaring the CMAKE_MINIMUM_VERSION we specify the minimum
version of cmake needed in order to run _AND_ how it should work.

In order to allow for changing the behavior of functionality
in cmake it offers a "policy" method which may be set to OLD
(to get the behavior it used to have) or NEW to get the new
behavior. By default setting CMAKE_MINIMUM_VERSION cause
all of the settings defined up to that version to NEW, and
each of the setting may then be set to OLD in order to force
the old behavior. The old behavior is typically deprecated
and may be removed in a future version.

To avoid getting "too far behind" on cmake policies (and
even more work the next time we want to upgrade) this patch
brings us up to 3.2.

Change-Id: I099bc4d74cd79cb3558e8735dcd664de1f694a35
Reviewed-on: http://review.couchbase.org/95415
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>