Cleanup: Disallow mix of compiler suites
Enforce that the C compiler is the "brand" as the C++
compiler, and then move all of the settings for each
compiler brand into the same file.

The C and C++ compiler is tightly coupled together and
it is easier to have to deal with a single file per
compiler suite rather than having to update multiple
files.

Change-Id: Ifc776df1ad3879de760df46e4825a3e968d78bd1
Reviewed-on: http://review.couchbase.org/95487
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>