MB-30852: Downgrade to RocksDB 5.12.5
I wanted to upgrade from 5.8 to the latest 5.14 version, but guys at
Facebook removed a compaction-callback that we need from version 5.13.
So, for now I upgrade to the latest 5.12.

Note that I opened an issue at
https://github.com/facebook/rocksdb/issues/4288.
There is a good chance that the callback we need will be re-introduced
in the short term.

Change-Id: Iafbd460cb2f9e3ca2cc6cac17e1806495943010e
Reviewed-on: http://review.couchbase.org/99410
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>