Add HeaderObjectLibrary.cmake
Adds add_header_object_library(): Compile a set of (C++) header files
into an Object library.

Motivation:

This isn't normally possible in CMake (any headers defined as
dependancies on a target are ignored in terms of what it built), but
in two specific situations it can be useful to have CMake actually
compile a header:

1. To verify that header have all required #include directives

2. To determine how costly a header is to compile.
   (This is the primary motivation for adding this function).

Usage:

    add_header_object_library(
        NAME <target_name>
        HEADERS <headers...>
    )

This will define an object library NAME, which consists of the
specified list of headers.

The object library isn't expected to be linked against - it exists
only for the above two reasons, so the headers get compiled. As such
the target will be excluded from the 'all' target - to build it one
must explicilty specify that target to be build.

Change-Id: I7fb70889304cbff0f516bd74ce3ce67572d9a0aa
Reviewed-on: http://review.couchbase.org/99800
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>