MB-30040: Use layout=system, address-model=64 for Boost
Under the windows build; boost's build system defaults to the
'versioned' layout which inserts a versioned path into name of the
directory include files are installed - e.g.

    include/boost-1.67/boost/

However this doesn't match *ix - and consequently where we expect to
locate Boost headers. Explicitly set the layout to 'system' which uses
the traditional layout we expect:

    include/boost/

This also requires that we only build one address model (64); as
otherwise there are path conflicts in installing the 32bit and 64bit
libs.

Given this changes our Boost cbdep package; bump version to 1.67-cb2.

Change-Id: I8ea7885f425800cb4064fb8e0c6cc04126df12f7
Reviewed-on: http://review.couchbase.org/100625
Tested-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>