MB-28835: Set library run path relative to binary location for curl
To avoid needing to set LD_LIBRARY_PATH, the build for the curl cbdep
has been configured to set the proper run path on Linux and macOS (this
change should not be needed on Windows).  This is done via setting the
LDFLAGS environment variable for linux, and using install_name_tool on
the libcurl library as well as the curl binary on macOS.

Becuase of difficulties with CMake in trying to accomplish this, along
with a desire to prep for the "cbdeps 2.0" conversion, the CMake scripts
were updated to use shell/batch scripts for the actual builds.

Change-Id: I9504bc09568422c8971a0d7571dc7eb3b5633625
Reviewed-on: http://review.couchbase.org/100946
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>